requirements.txt
flask>=2.0
prometheus-client>=0.14
requests>=2.28
python-dotenv>=1.0